implement condensed view for tumour and patient info
