+++ app/views/ReportView/components/PatientInformation/index.scss
@@ -1,0 +1,15 @@
+@import '../../../../styles/variables';
+
+.patient-information {
+    margin: 0 0 32px 0;
+
+    &__content {
+      padding: 0 16px;
+    }
+}
+
+.patient-information--print{
+    &__title {
+        margin: 0 0 8px;
+    }
+}
