+++ app/views/ReportView/components/ProbeSummary/index.tsx
@@ -1,31 +1,27 @@
 import React, {
-  useState, useEffect, useContext, useCallback, useMemo,
+  useState, useEffect, useContext, useCallback, useMemo, lazy,
 } from 'react';
 import {
   Typography,
-  IconButton,
-  Grid,
 } from '@mui/material';
-import EditIcon from '@mui/icons-material/Edit';
 import capitalize from 'lodash/capitalize';
 import api, { ApiCallSet } from '@/services/api';
 import snackbar from '@/services/SnackbarUtils';
 import DataTable from '@/components/DataTable';
-import ReportContext, { ReportType, PatientInformationType } from '@/context/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import useReport from '@/hooks/useReport';
 import ConfirmContext from '@/context/ConfirmContext';
-import ReadOnlyTextField from '@/components/ReadOnlyTextField';
-import { formatDate } from '@/utils/date';
 import SignatureCard, { SignatureType, SignatureUserType } from '@/components/SignatureCard';
 import PrintTable from '@/components/PrintTable';
 import TestInformation, { TestInformationType } from '@/components/TestInformation';
 import withLoading, { WithLoadingInjectedProps } from '@/hoc/WithLoading';
-import PatientEdit from '@/components/PatientEdit';
 import EventsEditDialog from '@/components/EventsEditDialog';
 import { sampleColumnDefs, eventsColumnDefs } from './columnDefs';
 import ProbeResultsType from './types.d';
 
 import './index.scss';
+
+const PatientInformation = lazy(() => import('../PatientInformation'));
 
 type ProbeSummaryProps = {
   loadedDispatch: (type: { type: string }) => void;
@@ -38,7 +34,7 @@ const ProbeSummary = ({
   isPrint,
   setIsLoading,
 }: ProbeSummaryProps): JSX.Element => {
-  const { report, setReport } = useContext(ReportContext);
+  const { report } = useContext(ReportContext);
   let { canEdit } = useReport();
   if (report.state === 'completed') {
     canEdit = false;
@@ -48,13 +44,8 @@ const ProbeSummary = ({
   const [testInformation, setTestInformation] = useState<TestInformationType | null>();
   const [signatures, setSignatures] = useState<SignatureType | null>();
   const [probeResults, setProbeResults] = useState<ProbeResultsType[] | null>();
-  const [patientInformation, setPatientInformation] = useState<{
-    label: string;
-    value: string | null;
-  }[] | null>();
   const [editData, setEditData] = useState();
 
-  const [showPatientEdit, setShowPatientEdit] = useState(false);
   const [showEventsDialog, setShowEventsDialog] = useState(false);
 
   const classNamePrefix = isPrint ? 'probe-summary--print' : 'probe-summary';
@@ -96,41 +87,6 @@ const ProbeSummary = ({
             });
           });
           setProbeResults(probeResultsData);
-
-          setPatientInformation([
-            {
-              label: 'Alternate ID',
-              value: report.alternateIdentifier,
-            },
-            {
-              label: 'Pediatric Patient IDs',
-              value: report.pediatricIds,
-            },
-            {
-              label: 'Report Date',
-              value: formatDate(report.createdAt),
-            },
-            {
-              label: 'Case Type',
-              value: report.patientInformation.caseType,
-            },
-            {
-              label: 'Physician',
-              value: report.patientInformation.physician,
-            },
-            {
-              label: 'Biopsy Name',
-              value: report.biopsyName,
-            },
-            {
-              label: 'Biopsy Details',
-              value: report.patientInformation.biopsySite,
-            },
-            {
-              label: 'Gender',
-              value: report.patientInformation.gender,
-            },
-          ]);
         } catch (err) {
           snackbar.error(`Network error: ${err}`);
         } finally {
@@ -145,57 +101,6 @@ const ProbeSummary = ({
     }
   }, [loadedDispatch, report, setIsLoading]);
 
-  const handlePatientEditClose = useCallback((
-    newPatientData: PatientInformationType,
-    newReportData: ReportType,
-  ) => {
-    setShowPatientEdit(false);
-
-    if (!newPatientData && !newReportData) {
-      return;
-    }
-
-    if (newReportData) {
-      setReport((oldReport) => ({ ...oldReport, ...newReportData }));
-    }
-
-    if (newPatientData) {
-      setPatientInformation([
-        {
-          label: 'Alternate ID',
-          value: newReportData ? newReportData.alternateIdentifier : report.alternateIdentifier,
-        },
-        {
-          label: 'Pediatric Patient IDs',
-          value: newReportData ? newReportData.pediatricIds : report.pediatricIds,
-        },
-        {
-          label: 'Report Date',
-          value: formatDate(report.createdAt),
-        },
-        {
-          label: 'Case Type',
-          value: newPatientData ? newPatientData.caseType : report.patientInformation.caseType,
-        },
-        {
-          label: 'Physician',
-          value: newPatientData ? newPatientData.physician : report.patientInformation.physician,
-        },
-        {
-          label: 'Biopsy Name',
-          value: newReportData ? newReportData.biopsyName : report.biopsyName,
-        },
-        {
-          label: 'Biopsy Details',
-          value: newPatientData ? newPatientData.biopsySite : report.patientInformation.biopsySite,
-        },
-        {
-          label: 'Gender',
-          value: newPatientData ? newPatientData.gender : report.patientInformation.gender,
-        },
-      ]);
-    }
-  }, [report, setReport]);
 
   const handleSign = useCallback((signed: boolean, role: SignatureUserType) => {
     let cancelled;
@@ -297,41 +202,12 @@ const ProbeSummary = ({
     <div className={classNamePrefix}>
       {!isLoading && (
         <>
-          {report && patientInformation && (
-            <div className={`${classNamePrefix}__patient-information`}>
-              <div className={`${classNamePrefix}__patient-information-title`}>
-                <Typography variant="h3" display="inline">
-                  Patient Information
-                  {canEdit && !isPrint && (
-                    <>
-                      <IconButton onClick={() => setShowPatientEdit(true)} size="large">
-                        <EditIcon />
-                      </IconButton>
-                      <PatientEdit
-                        patientInformation={report.patientInformation}
-                        report={report}
-                        isOpen={Boolean(showPatientEdit)}
-                        onClose={handlePatientEditClose}
-                      />
-                    </>
-                  )}
-                </Typography>
-              </div>
-              <Grid
-                alignItems="flex-end"
-                container
-                spacing={3}
-                className={`${classNamePrefix}__patient-information-content`}
-              >
-                {patientInformation.map(({ label, value }) => (
-                  <Grid key={label} item>
-                    <ReadOnlyTextField label={label}>
-                      {value}
-                    </ReadOnlyTextField>
-                  </Grid>
-                ))}
-              </Grid>
-            </div>
+          {report && (
+            <PatientInformation
+              canEdit={canEdit}
+              isPrint={isPrint}
+              loadedDispatch={loadedDispatch}
+            />
           )}
           {report && report.sampleInfo && (
             <div className={`${classNamePrefix}__sample-information`}>
