+++ app/views/ReportView/components/RapidSummary/index.tsx
@@ -1,31 +1,23 @@
 import React, {
-  useState, useEffect, useContext, useCallback, useMemo,
+  useState, useEffect, useContext, useCallback, useMemo, lazy,
 } from 'react';
 import {
   Typography,
-  IconButton,
-  Grid,
 } from '@mui/material';
-import EditIcon from '@mui/icons-material/Edit';
 
 import api, { ApiCallSet } from '@/services/api';
 import snackbar from '@/services/SnackbarUtils';
 import DataTable from '@/components/DataTable';
-import ReportContext, { ReportType, PatientInformationType } from '@/context/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import useReport from '@/hooks/useReport';
 import ConfirmContext from '@/context/ConfirmContext';
-import ReadOnlyTextField from '@/components/ReadOnlyTextField';
-import { formatDate } from '@/utils/date';
 import SignatureCard, { SignatureType, SignatureUserType } from '@/components/SignatureCard';
 import PrintTable from '@/components/PrintTable';
 import withLoading, { WithLoadingInjectedProps } from '@/hoc/WithLoading';
-import PatientEdit from '@/components/PatientEdit';
 import capitalize from 'lodash/capitalize';
 import orderBy from 'lodash/orderBy';
 
 import './index.scss';
-import TumourSummaryEdit from '@/components/TumourSummaryEdit';
-import DescriptionList from '@/components/DescriptionList';
 import {
   KbMatchType, TumourSummaryType, ImmuneType, MutationBurdenType, MicrobialType, TmburType,
 } from '@/common';
@@ -36,6 +28,9 @@ import {
 import { VariantEditDialog, FIELDS } from './components/VariantEditDialog';
 import { RapidVariantType } from './types';
 import { getVariantRelevanceDict } from './utils';
+
+const PatientInformation = lazy(() => import('../PatientInformation'));
+const TumourSummary = lazy(() => import('../TumourSummary'));
 
 const splitIprEvidenceLevels = (kbMatches: KbMatchType[]) => {
   const iprRelevanceDict = {};
@@ -131,7 +126,7 @@ const RapidSummary = ({
   isPrint = false,
   setIsLoading,
 }: RapidSummaryProps): JSX.Element => {
-  const { report, setReport } = useContext(ReportContext);
+  const { report } = useContext(ReportContext);
   const { setIsSigned } = useContext(ConfirmContext);
   let { canEdit } = useReport();
   if (report.state === 'completed') {
@@ -142,10 +137,6 @@ const RapidSummary = ({
   const [therapeuticAssociationResults, setTherapeuticAssociationResults] = useState<RapidVariantType[] | null>();
   const [cancerRelevanceResults, setCancerRelevanceResults] = useState<RapidVariantType[] | null>();
   const [unknownSignificanceResults, setUnknownSignificanceResults] = useState<RapidVariantType[] | null>();
-  const [patientInformation, setPatientInformation] = useState<{
-    label: string;
-    value: string | null;
-  }[] | null>();
   const [tumourSummary, setTumourSummary] = useState<TumourSummaryType[]>();
   const [primaryBurden, setPrimaryBurden] = useState<MutationBurdenType>();
   const [tmburMutBur, setTmburMutBur] = useState<TmburType>();
@@ -153,8 +144,6 @@ const RapidSummary = ({
   const [microbial, setMicrobial] = useState<MicrobialType[]>();
   const [editData, setEditData] = useState();
 
-  const [showPatientEdit, setShowPatientEdit] = useState(false);
-  const [showTumourSummaryEdit, setShowTumourSummaryEdit] = useState(false);
   const [showMatchedTumourEditDialog, setShowMatchedTumourEditDialog] = useState(false);
   const [showCancerRelevanceEventsDialog, setShowCancerRelevanceEventsDialog] = useState(false);
 
@@ -245,41 +234,6 @@ const RapidSummary = ({
           } else if (!isPrint) {
             snackbar.error(microbialResp.reason?.content?.error?.message);
           }
-
-          setPatientInformation([
-            {
-              label: 'Alternate ID',
-              value: report.alternateIdentifier,
-            },
-            {
-              label: 'Report Date',
-              value: formatDate(report.createdAt),
-            },
-            {
-              label: 'Case Type',
-              value: report.patientInformation.caseType,
-            },
-            {
-              label: 'Physician',
-              value: report.patientInformation.physician,
-            },
-            {
-              label: 'Biopsy Name',
-              value: report.biopsyName,
-            },
-            {
-              label: 'Biopsy Details',
-              value: report.patientInformation.biopsySite,
-            },
-            {
-              label: 'Gender',
-              value: report.patientInformation.gender,
-            },
-            {
-              label: 'Tumour type for matching',
-              value: report.kbDiseaseMatch,
-            },
-          ]);
         } catch (err) {
           snackbar.error(`Unknown error: ${err}`);
         } finally {
@@ -390,58 +344,6 @@ const RapidSummary = ({
   }, [microbial, primaryBurden, tmburMutBur, report.m1m2Score, report.sampleInfo, report.tumourContent, tCellCd8?.percentile, tCellCd8?.score, report.captiv8Score,
     tCellCd8?.percentileHidden, tCellCd8, tCellCd8?.pedsScoreComment, tmburMutBur?.adjustedTmb, tmburMutBur?.tmbHidden, tCellCd8?.pedsScore, tCellCd8?.pedsPercentile]);
 
-  const handlePatientEditClose = useCallback((
-    newPatientData: PatientInformationType,
-    newReportData: ReportType,
-  ) => {
-    setShowPatientEdit(false);
-
-    if (!newPatientData && !newReportData) {
-      return;
-    }
-
-    if (newReportData) {
-      setReport((oldReport) => ({ ...oldReport, ...newReportData }));
-    }
-
-    if (newPatientData) {
-      setPatientInformation([
-        {
-          label: 'Alternate ID',
-          value: newReportData ? newReportData.alternateIdentifier : report.alternateIdentifier,
-        },
-        {
-          label: 'Report Date',
-          value: formatDate(report.createdAt),
-        },
-        {
-          label: 'Case Type',
-          value: newPatientData ? newPatientData.caseType : report.patientInformation.caseType,
-        },
-        {
-          label: 'Physician',
-          value: newPatientData ? newPatientData.physician : report.patientInformation.physician,
-        },
-        {
-          label: 'Biopsy Name',
-          value: newReportData ? newReportData.biopsyName : report.biopsyName,
-        },
-        {
-          label: 'Biopsy Details',
-          value: newPatientData ? newPatientData.biopsySite : report.patientInformation.biopsySite,
-        },
-        {
-          label: 'Gender',
-          value: newPatientData ? newPatientData.gender : report.patientInformation.gender,
-        },
-        {
-          label: 'Tumour type for matching',
-          value: newReportData ? newReportData.kbDiseaseMatch : report.kbDiseaseMatch,
-        },
-      ]);
-    }
-  }, [report, setReport]);
-
   const handleSign = useCallback((signed: boolean, role: SignatureUserType) => {
     let cancelled;
     const sign = async (s: boolean, r: SignatureUserType) => {
@@ -459,73 +361,6 @@ const RapidSummary = ({
     sign(signed, role);
     return function cleanup() { cancelled = true; };
   }, [report.ident, setIsSigned]);
-
-  const handleTumourSummaryEditClose = useCallback((
-    isSaved,
-    newMicrobialData,
-    newReportData,
-    newTCellCd8Data: ImmuneType,
-    newMutationBurdenData: MutationBurdenType,
-    newTmBurMutBurData,
-  ) => {
-    setShowTumourSummaryEdit(false);
-
-    if (!isSaved || (!newMicrobialData && !newReportData && !newTCellCd8Data && !newMutationBurdenData && !newTmBurMutBurData)) {
-      return;
-    }
-
-    if (newMicrobialData) {
-      setMicrobial(newMicrobialData);
-    }
-
-    if (newReportData) {
-      setReport(newReportData);
-    }
-
-    if (newTCellCd8Data) {
-      setTCellCd8(newTCellCd8Data);
-    }
-
-    if (newMutationBurdenData) {
-      setPrimaryBurden(newMutationBurdenData);
-    }
-
-    if (newTmBurMutBurData) {
-      setTmburMutBur(newTmBurMutBurData);
-    }
-  }, [setReport]);
-
-  let tumourSummarySection = null;
-  if (tumourSummary) {
-    tumourSummarySection = (
-      <>
-        <div className="rapid-summary__tumour-summary-title">
-          <Typography variant="h3">
-            Tumour Summary
-            {canEdit && !isPrint && (
-              <>
-                <IconButton onClick={() => setShowTumourSummaryEdit(true)} size="large">
-                  <EditIcon />
-                </IconButton>
-                <TumourSummaryEdit
-                  microbial={microbial}
-                  report={report}
-                  tCellCd8={tCellCd8}
-                  mutationBurden={primaryBurden}
-                  tmburMutBur={tmburMutBur}
-                  isOpen={showTumourSummaryEdit}
-                  onClose={handleTumourSummaryEditClose}
-                />
-              </>
-            )}
-          </Typography>
-        </div>
-        <div className="rapid-summary__tumour-summary-content">
-          <DescriptionList entries={tumourSummary} />
-        </div>
-      </>
-    );
-  }
 
   const handleMatchedTumourEditStart = useCallback((rowData) => {
     setShowMatchedTumourEditDialog(true);
@@ -756,46 +591,19 @@ const RapidSummary = ({
     <div className={`rapid-summary${isPrint ? '--print' : ''}`}>
       {!isLoading && (
         <>
-          {report && patientInformation && (
-            <div className="rapid-summary__patient-information">
-              <div className="rapid-summary__patient-information-title">
-                <Typography variant="h3" display="inline">
-                  Patient Information
-                  {canEdit && !isPrint && (
-                    <>
-                      <IconButton onClick={() => setShowPatientEdit(true)} size="large">
-                        <EditIcon />
-                      </IconButton>
-                      <PatientEdit
-                        patientInformation={report.patientInformation}
-                        report={report}
-                        isOpen={Boolean(showPatientEdit)}
-                        onClose={handlePatientEditClose}
-                      />
-                    </>
-                  )}
-                </Typography>
-              </div>
-              <Grid
-                alignItems="flex-end"
-                container
-                spacing={3}
-                className="rapid-summary__patient-information-content"
-              >
-                {patientInformation.map(({ label, value }) => (
-                  <Grid key={label} item>
-                    <ReadOnlyTextField label={label}>
-                      {value}
-                    </ReadOnlyTextField>
-                  </Grid>
-                ))}
-              </Grid>
-            </div>
+          {report && (
+            <PatientInformation
+              canEdit={canEdit}
+              isPrint={isPrint}
+              loadedDispatch={loadedDispatch}
+            />
           )}
           {report && tumourSummary && (
-            <div className="rapid-summary__tumour-summary">
-              {tumourSummarySection}
-            </div>
+            <TumourSummary
+              canEdit={canEdit}
+              isPrint={isPrint}
+              tumourSummary={tumourSummary}
+            />
           )}
           {report && therapeuticAssociationResults && (
             <div className="rapid-summary__events">
